% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ConvexNMF_ADMM}
\alias{ConvexNMF_ADMM}
\title{Convex NMF for mix-signed matricies using ADMM}
\usage{
ConvexNMF_ADMM(V, k, cluster, verbose = 0L, maxiter = 100L,
  fixediter = 0L, rho = 3, tol = 0.001)
}
\arguments{
\item{V}{A (m x n) vector of data matrix with m features and n units}

\item{k}{An integer to specify desired rank of the low rank factorization result}

\item{cluster}{A (k x 1) vector of integer 0~(k-1) indicating initial cluster assignment}

\item{verbose}{A 0/1 integer (default 0). If 1: output the relative error after each iteration}

\item{maxiter}{An integer (default 100) for the maximal number of iterations}

\item{fixediter}{An integer (optional) for a fixed number of iterations}

\item{rho}{Penalty parameter in ADMM (default 3)}

\item{tol}{Tolerance parameter to declare convergence}
}
\value{
A list containing the following values
* W     : A (n x k) non-negative matrix containing W in X~XWG
* G     : A (k x n) non-negative matrix containing G in X~XWG
* Error : A vector containing the relative distance of this approximation in each iteration
}
\description{
This function implements a ADMM for NMF taking a single matrix
of any shape and potentially mix-signed.
X~XWG with W & G non-negative.
}
